% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_column.R
\name{split_column}
\alias{split_column}
\title{Split column into new columns}
\usage{
split_column(column, split)
}
\arguments{
\item{column}{character vector, column to split}

\item{split}{character, how to split values}
}
\description{
Split a column with a given delimiter, and output into a data.frame. The object
can be bound to the input data.frame
}
\examples{

tmp <- data.frame(cat = c('ABC', 'BCA', 'CAB', '', 'AXC', '12345', NA, '1'))
split_column(tmp$cat, split = 'A')
split_column(tmp$cat, split = '')

}
